Switched download button from list to expanded fragment
