+++ app/src/main/res/menu/menu_expanded_fragment.xml
@@ -1,0 +1,10 @@
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools" >
+    <item
+        android:id="@+id/item_download"
+        android:title="Baixar grade"
+        android:icon="@drawable/ic_download"
+        android:iconTint="#FFF"
+        app:showAsAction="always" />
+</menu>
